\name{BartMixVs-package}
\alias{BartMixVs-package}
\alias{BartMixVs}
\docType{package}
\title{
  Varibale Selection Using Bayesian Additive Regression Trees
}
\description{
  This package provides implementations of existing BART-based variable selection approaches.
}
\details{
  Bayesian additive regression trees (BART) provides flexible nonparametric modeling of mixed-type predictors for continuous and binary responses. This package is built upon CRAN R package BART, version 2.7 (https://github.com/cran/BART). It implements the three proposed variable selection approaches of the paper, Luo, C and Daniels, MJ (2022), "Variable Selection Using Bayesian Additive Regreesion Trees", and other three existing BART-based variable selection approaches.
}
\author{
Chuji Luo <cjluo@ufl.edu>, Michael J. Daniels <daniels@ufl.edu>

Maintainer: Chuji Luo <cjluo@ufl.edu>
}
\references{
  LUO, C and DANIELS, MJ (2022). Variable Selection Using Bayesian Additive Regression Trees.
  
  BLEICH, J., KAPELNER, A., GEORGE, E. I. and JENSEN, S. T. (2014). Variable selection for BART: an application to gene reg- ulation. Ann. Appl. Stat. 8 1750–1781.
  
  LINERO, A. R. (2018). Bayesian regression trees for high-dimensional prediction and variable selection. J. Amer. Statist. Assoc. 113 626– 636.
  
  LIU, Y., ROCKOVÁ, V. and WANG, Y. (2021). Variable selection with
ABC Bayesian forests. J. R. Stat. Soc. Ser. B. Stat. Methodol. 83
453–481.

  SPARAPANI, R., SPANBAUER, C. and MCCULLOCH, R. (2021).
Nonparametric machine learning and efficient computation with bayesian additive regression trees: the BART R package. J. Stat. Softw. 97 1–66.
}
\keyword{BART, Variable Selection}
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
